+++ replacer-backend/src/main/java/es/bvalero/replacer/review/dto/ReviewSectionDto.java
@@ -19,4 +19,8 @@ public class ReviewSectionDto {
     @NonNull
     @NotNull
     String title;
+
+    @Schema(description = "Offset of the section with the page content", required = true, example = "1014")
+    @NotNull
+    int offset;
 }